Make image backend import from oldagg rather than the other way around.
